Move the logic to generate/prepare a deploy into deploy.
This moves all the Docker image building into the deploy module, meaning
the external CLI (or server) need only pass a list of (git) directories.
