+++ kubetools_client/cli/deploy.py
@@ -4,13 +4,13 @@ import os
 import click
 
 from kubetools_client.cli import cli_bootstrap
-from kubetools_client.config import load_kubetools_config
 from kubetools_client.deploy import (
     execute_cleanup,
     execute_deploy,
     execute_remove,
     execute_restart,
     get_cleanup_objects,
+    get_deploy_objects,
     get_remove_objects,
     get_restart_objects,
     log_cleanup_changes,
@@ -19,13 +19,7 @@ from kubetools_client.deploy import (
     log_restart_changes,
 )
 from kubetools_client.deploy.build import Build
-from kubetools_client.deploy.image import ensure_docker_images
 from kubetools_client.deploy.kubernetes.api import get_object_name
-from kubetools_client.deploy.kubernetes.config import (
-    generate_kubernetes_configs_for_project,
-)
-from kubetools_client.deploy.util import run_shell_command
-from kubetools_client.exceptions import KubeBuildError
 
 
 def _dry_deploy_object_loop(object_type, objects):
@@ -55,37 +49,6 @@ def _dry_deploy_loop(build, services, deployments, jobs):
     ):
         if objects:
             _dry_deploy_object_loop(object_type, objects)
-
-
-def _get_git_info(app_dir):
-    git_annotations = {}
-
-    commit_hash = run_shell_command(
-        'git', 'rev-parse', '--short=7', 'HEAD',
-        cwd=app_dir,
-    ).strip().decode()
-    git_annotations['kubetools/git_commit'] = commit_hash
-
-    branch_name = run_shell_command(
-        'git', 'rev-parse', '--abbrev-ref', 'HEAD',
-        cwd=app_dir,
-    ).strip().decode()
-
-    if branch_name != 'HEAD':
-        git_annotations['kubetools/git_branch'] = branch_name
-
-    try:
-        git_tag = run_shell_command(
-            'git', 'tag', '--points-at', commit_hash,
-            cwd=app_dir,
-        ).strip().decode()
-    except KubeBuildError:
-        pass
-    else:
-        if git_tag:
-            git_annotations['kubetools/git_tag'] = git_tag
-
-    return commit_hash, git_annotations
 
 
 @cli_bootstrap.command(help_priority=0)
@@ -131,56 +94,17 @@ def deploy(ctx, dry, replicas, registry, yes, namespace, app_dirs):
         namespace=namespace,
     )
 
-    all_services = []
-    all_deployments = []
-    all_jobs = []
-
-    for app_dir in app_dirs:
-        envvars = {
-            'KUBE_ENV': build.env,
-            'KUBE_NAMESPACE': build.namespace,
-        }
-
-        annotations = {
-            'kubetools/env': build.env,
-            'kubetools/namespace': build.namespace,
-        }
-
-        if os.path.exists(os.path.join(app_dir, '.git')):
-            commit_hash, git_annotations = _get_git_info(app_dir)
-            annotations.update(git_annotations)
-        else:
-            raise click.BadParameter(f'{app_dir} is not a valid git repository!')
-
-        kubetools_config = load_kubetools_config(
-            app_dir,
-            env=build.env,
-            namespace=build.namespace,
-        )
-
-        context_to_image = ensure_docker_images(
-            kubetools_config, build, app_dir,
-            commit_hash=commit_hash,
-            default_registry=registry,
-        )
-
-        services, deployments, jobs = generate_kubernetes_configs_for_project(
-            kubetools_config,
-            envvars=envvars,
-            context_name_to_image=context_to_image,
-            base_annotations=annotations,
-            replicas=replicas,
-        )
-
-        all_services.extend(services)
-        all_deployments.extend(deployments)
-        all_jobs.extend(jobs)
+    services, deployments, jobs = get_deploy_objects(
+        build, app_dirs,
+        replicas=replicas,
+        default_registry=registry,
+    )
 
     if dry:
-        return _dry_deploy_loop(build, all_services, all_deployments, all_jobs)
+        return _dry_deploy_loop(build, services, deployments, jobs)
 
     log_deploy_changes(
-        build, all_services, all_deployments, all_jobs,
+        build, services, deployments, jobs,
         message='Executing changes:' if yes else 'Proposed changes:',
         name_formatter=lambda name: click.style(name, bold=True),
     )
@@ -194,9 +118,9 @@ def deploy(ctx, dry, replicas, registry, yes, namespace, app_dirs):
 
     execute_deploy(
         build,
-        all_services,
-        all_deployments,
-        all_jobs,
+        services,
+        deployments,
+        jobs,
     )
